+++ packages/@eventual/testing/src/clients/timer-client.ts
@@ -1,11 +1,11 @@
+import { assertNever } from "@eventual/core";
 import {
-  assertNever,
   computeScheduleDate,
   isActivityHeartbeatMonitorRequest,
   isTimerScheduleEventRequest,
   TimerClient,
   TimerRequest,
-} from "@eventual/core";
+} from "@eventual/runtime-core";
 import { TimeConnector } from "../environment.js";
 
 export class TestTimerClient extends TimerClient {