+++ packages/@eventual/timeline/src/App.tsx
@@ -1,12 +1,15 @@
+import { HttpServiceClient } from "@eventual/client";
+import { decodeExecutionId, WorkflowStarted } from "@eventual/core";
 import { useQuery } from "@tanstack/react-query";
-import { Buffer } from "buffer";
-import ky from "ky";
 import { ReactNode } from "react";
-import { ActivityList } from "./components/activity-list/activity-list.js";
+import { aggregateEvents } from "./activity.js";
 import styles from "./App.module.css";
-import { TimelineActivity } from "./activity.js";
+import { ActivityList } from "./components/activity-list/activity-list.js";
 import { Timeline } from "./components/timeline/timeline.js";
-import { WorkflowStarted } from "@eventual/core";
+
+const serviceClient = new HttpServiceClient({
+  serviceUrl: `${window.location.origin}/api`,
+});
 
 function Layout({
   start,
@@ -24,9 +27,7 @@ function Layout({
   if (!executionId) {
     return <div>No execution id in path!</div>;
   }
-  const decodedExecutionId = Buffer.from(executionId, "base64").toString(
-    "utf-8"
-  );
+  const decodedExecutionId = decodeExecutionId(executionId);
   return (
     <main className={styles.layout}>
       <div style={{ textAlign: "center" }}>
@@ -64,17 +65,24 @@ function Layout({
 }
 
 function App() {
-  const { data: timeline, isLoading } = useQuery(
-    ["events"],
-    () => {
+  const { data: timeline, isLoading } = useQuery({
+    queryKey: ["events"],
+    queryFn: async () => {
       const executionId = window.location.href.split("/").at(-1);
-      return ky(`/api/timeline/${executionId}`).json<{
-        start: WorkflowStarted;
-        activities: TimelineActivity[];
-      }>();
+      const history = await serviceClient.getExecutionWorkflowHistory(
+        decodeExecutionId(executionId!)
+      );
+      const { activities, start } = aggregateEvents(history.events);
+      return {
+        activities,
+        start,
+      };
     },
-    { refetchInterval: 5000 }
-  );
+    onError: (err) => {
+      console.log(err);
+    },
+    refetchInterval: 5000,
+  });
 
   if (isLoading) {
     return (